Core functionality of Vue.
Player and Monster game, including Health bar, with abilities such as - Attack, Special Attack, Heal and Surrender. Battle log of what happens during the battle, including every move the person choose. Using core functionality - methods, computed, watch, data, interpolation, conditional statements, binding. Two separate functions of formulas for AttackedHealth and Healing, they are used below in the methods to calculate the random damage the player deals to the monster and monster to the player. Every time the player attacks the monster, the monster deals random damage. Even when the players is healing himself. Player can use special attack every 3 rounds, because it becomes unkillable. Data log is added to an array with the function unshift, to be send beginning of the array. The losers is the one that gets to zero health at the Health bar. There is a chance for Draw! Health bar is done by using dynamic styling. The watchers, watches for a winner.