Resource App done!
Proper styling with two sections - Stored resources and Add resources, good structure of the folders and the elements, using components features like a providing , injection , props ,emits events ,slot. You can add resource and go to the provided link you've entered. If you want the you can remove it. If you do not enter a character in one of the 3 inputs you will get a dialog with error, there you will have a paragraph and a button okay. You can click the button okay to close it or you can click everywhere except on the dialog. That is one little project using good features.