Validating the form.
Using variables that are - empty or null,2 methods one to validate the form, in which I use if checks, checking every object if the value is empty. Once the input field have an symbol the indication of error is gone. The second method is to clear the errors once I type a symbol in the input. With the help of @blur it checks once there is a symbol in the input.