Added Error Handling in The Requests
Async methond loadRequests, which has try catch error , set isLoading first to true ,after the dispatching it's again set to false isLoading to stop the loading spinner. An error dialog occures if there is an error, can be close by the button or the click out of the dialog, fetching data is done too.