Attach Token
Mutations helped me to attach access the token whenever I need it using ?auth= + token name in the fetch url. Now when a racer sign up and the log in there is no more errors in the requests, now they are empty. Everything is on the firebase now. Working on the UI and some other implementations.