Improved Interface
Added getter isAuthenticated() and return the authenticates racers using !! to convert to 1 token. In RacerList, I added another one baseButton which leads to auth if the racers is not logged. The old base button is improved too. Now if the racers is logged in, is not a racer and is not loading the register as a racer button appears. Now I'm going to improve more features. Working on them