Navigation guard added
FIrstly, after log in and log out the racer is redirected to the racers pages. Navigation guard is added in the router.js  and have some if statements, if the racer is not authenticated and the link requires authentication it should redirect to /auth  , else if the link doesn't requires authentication and the racers is authenticated it will redirect to racers and the last else will continue, that's the navigation guard.