New Project Started.
MobileO is a SPA - It includes, router and later Vuex. The main point is a website for cars sales. Including all forms with filters and many other options. To run this in your computer you first need to type in the terminal "npm install" to add the folder with modules. After that you can run it.