Usage of props in the shortest form
Simple Application, with One component FriendContact, that contains id,name,email and phone number. In the main component App added two component and to see them in the server with different data I needed to use props. In the component file I added props that include name,email and phone number
and interpolate them in the FriendContact component. Then passed the needed props in the app File. Now there are two Components with different data.