New Component Added
TrainersItem, added where all the main information that is need is visible. Using props to send data from parent TrainersList to TrainersItem. In TrainersItem I created 3 computed properties, the first one is the full name, the second is link to the contact page and the third is link to details page, using the id of the current  gymCoach. In TrainersList register the TrainerItem component and used it with v-for that iterates in the getter of all trainers. V-binding the first,last name , the rate and the areas of expertise. Key is the trainer id.