New Base component
BaseButton added, includes an button, that contains slot and a router link, that contains slot too. 3 props that are validated, mode, link, to. For the button I check if the link is true, if it's not then it's a button.  to property that by default contains '/' and is bind to the router link to go to the current path. For styling class is bind to mode.
In the TrainersItem I replaced router links with base buttons with attribute link and :to bind to the computed properties. In TrainersList replaced the link and the button with base buttons. The refresh button is using the attribute mode to outline for the styling. For the router is the same like in TrainersItem link and to the path,that I want