TrainersDetails Component Done!
HTML - 3 sections and wrapped with 3 base cards, computed properties for all the information. One base button , that is a link. SCRIPT - Making the props true in the router.js. Passing the Id prop from router to TrainerDetails. Making one variable that is null and on this variable got the trainer that trainer.id equals to the passed Id from router. Then Making all the computed properties and interpolate them.