New component added.
Filter done! New component TrainersFilter added, to keep the TrainersList component clean as possible.
HTML - The code is wrap in base-card, it has 3 span tags, with 3 inputs that are type checkbox and labels. Every input runs method change. If I uncheck one of the three, the method is activated.
SCRIPT - In data I added filters object, that contains the 3 areas of expertise. They are set to true, because they are checked in the inputs. One method SetFilters() that receives in default event object. I get the inputId , that one is checked and update the existing filters by copying them with the spread operator and setting the Id to active.
Then existing filters are set to the updated ones. Well after that I emit change-filter and pass the updatedFilters to the Parent component. Created emits.
In TrainersList I import the TrainersFilter component and register it. Then I call the component with @change-filter emit with method setFilters.  The methods contains as value updatedFilters that are passed through the emit. In Data I create activefilters object with all the 3 areas. In the method I set activeFilters to updatedFilters. In the computed properties I change filteredTrainer computed property. Getting the the from the store is saved in a constant and that const is used to filter through trainers that are equals to the areas.
Two checks of if it's contains that area and if activeFilters are the selected area. 3 Checks total.