TrainersForm Added and Now you can Register as a Gym Coach
I added new component TrainersForm to keep the code as clean as possible. Using two-way data binding with v-model, data properties for all the inputs and the textarea. I created one method submitForm, that have an formData object, of all the data properties. I emit this method to TrainersRegistration, to get the vuex Data from there. At the top in the form i used @submit.prevent ,to prevent from resetting the form. In TrainersRegister I imported the component and register it. Then using save data in the component and create new method sendForm that contains the data. Next in mutations.js I created registerTrainer with state and payload. then push the formData 'Payload' in the trainers array. In actions I created object of formData , that contains all the data properties of a trainer. Then I commit the registerTrainer and the data. In TrainersRegister then I dispatched the data and the registerTrainer action.
Then I used router.replace to send me to the trainers path, immediately after I registered as a trainer.