Registering Coach Validate
If a Coach is once registered, it can not click register as coach button again.
In the index.js store I harcode the userId
then wrote getters that return the userId.
In Trainers/Getters.js I wrote IsTrainer() method that gets the trainers and the userId. Using the method some , to get the trainers that fulfill the requirements. In TrainersList I get the IsTrainers from the store/getters.