Validation form done!
clearValidity method added to see if the inputs are filled or not. Using the [input] inside the method and setting the isValid to true. In the input fields calling the @blur function and then the clearValiditiy with it's input name. If the form is submitted, but has some error when typing new words in the inputs the errors will go away. Thanks to the @blur function.