Contact Form Added
Contact form added, One input field and one textarea with basebutton.
data properties, two-way data binding with v-model and one boolean variable that is set to true by default.
submitForm method which checks if the input for the email is empty if not contain an @ symbol, or if the message is empty. Using @submit.prevent on the form to prevent from reloading the page.