Requests are seen now.
New component added, using props - email and message. Using computed properties to go directly to the mail.
Two getters added requests,getting all the requests. HasRequests added to validate requests that it's length is greater than zero. In RequestsReceived I registerd the RequestItem and use it with v-for in requestsReceived. Binding the email to userEmail and message to message.