Filtered requests and fetch coach data.
Filtering requests added by getting the id from rootGetters. return the state by using the filter function.

In hasRequests, it gets the data from the getters.

In Actions making the Id constant and creating new fetch with put method and convert the data to JSON formatv

Checking the response if it's not ok and commit object with the copied data and the Id.