Fetching Trainer/Coach Data.
GET HTTP Request.
New Mutation added setTrainers.
In Actions.js new method loadTrainers is added getting all the trainers and checking if the response is not ok.

Creating an array of trainers and looping through responseData where I create an object of trainer with key and responseData props.
Pushing the trainer object in the trainers array and commit the setTrainers mutations with the array trainers.

In TrainersList I added the lifecycle hook created in which I load the loadTrainers Method.
It dispatches loadTrainers from the actions.js file, binding the refresh button to use the loadTrainers method.
Working on adding a BaseSpinner while loading