HTTP Sending requests done!
Now the requests are saved in the database and everything run's fine. Using the base dialog component in RequestsReceived and the base spinner component. try catch on the async method loadRequests if there are any errors. handleError method, setting the error back to null, for closing the base dialog.
In created lifecycle hook , I call the loadRequests method. Removed the id from the info that is saved in FireBase, by removing it in the actions.js object - requestsData. Added new mutation setRequests, that sets the requests to the payload.