New component PetFilter Added.
Filtering by Type of the animal: Dog, Cat.
In PetsFilter, new data object - filters that includes the pet type and is set to boolean true. New Method setFilters with attribute event added. Firstly I get the id, then the checked ones. Then I copy the filters with the spread operator and set the inputId to Checked.
Set current filters to the updatedfilters.
Finally emit with the updatedFilters object. In PetsList component I use the change-filter emit with setFilter method. Computed property filteredPets I change it to filter through the type and the activeFilters object that I create earlier. If the validation is filtered it retuns a true, else false. Lastly the method for the emit is setFilter as I said with attribute updatedFilters that is passed through the emit. Setting the acriveFilters to the updateFilters and now the Filter Method is done.