RegisterForm for pets added.
In <div> I added one Label and one Input for every data property, except that for description is textarea. Using Two-Way Data Binding to get the current answers: with V-model="".
Using .trim, to remove all the space that is left. For the type of pet you have two checkboxes and for the type of the breed again you have to checkboxes. Then create one method which contains an object of formData to log it in the console, want to see what is the result. Nothing new till now. The interesting part comes now, and that's the uploadImage function. It get's now the file and print it in the console, for that I'm using two methods, the first one openImageUpload get's the 'petImage' attribute. handleImageUpload(event) gets the file that I have selected and in the image I store the file in the image data property.  Under the label there is one input that there is the choose file button. Needs desing and more work to see if this uploadImage function actually works as it needs to.