Added Id, to the store and more...
Added userId to the store, harcoded for now, Getter userId with attribute state.
Return userId.
In actions replace the 'p4' Id , with the one that I created in the store. In the separate pets > getters.js I wrote new Getter function, that gets the pets which are filtered through some function.