Messages added.
Actions, getters, index and mutations added. Firstly imported the messageModule inside the store and registered it there.
In messages index.js I created new messages Array.  In actions I added contactPets method. that commits the object of data. New mutation added where the payload is pushed in the array of messages. At the end I dispatch the data in PetsContact component and after dispatching it I replace it with the link that I want to go after I send the data.
The messages cannot be seen now, only can be sent.