Message form done.
Now messages can be seen after somebody contact the owner of the pet.
I created RequestItem, that has props - email,messages and phoneNumber. Created one computed property that is a link to the email. Small errors adjustments.  Imported the requestItem  in petRequests. Using v-for looping through the messageRequests. Now if there are no messages, the  message in h3 is going to appear.