Fetch data
In getters.js, I created const of the userId.
Then return the state with filtered Id's.
In hasMessages, I return only the messages from the getters. Helps for getting the current Id of the pet. Now in Pets/actions.js I created one more const and get the userId from the rootGetters. I made the method asynchronous and store the fetch in the const response. The method type is 'PUT' and transfer the formData into JSON format. Checking if the response is not ok. At the end commit the method with all it's data and the userId.