Get HTTP Request: Fetch Data
In pets/action.js I created new asynchronous method that fetch data and loops through array of responseData. Push the pet in the pets array and commit the changes. In PetsFilter I added new BaseButton: Refresh and I created new method loadPets, that dispatch data from loadPets method in the action.js. In the created lifecycle hook I call the loadPets method and in the refresh button with @click call it too. Now the page can be refresh and the data that is going to be displayed is the real data  not the array that is in index.js in pets.