HTTP request
Requests are saved in the database and they are separated. Every requests has an Id and will be shown to the pet's owner that id matches to the one in the requests. New mutation created. In the getters instead of petId I wroted it inside the petCoach petSId. 2 methods that are asynchronous . ContactPets method is sending request to the firebase and fetchMessages is getting the data from the Firebase. In PetsRequest component added Spinner and error handling,everything works fine.