Adding UserAuth
Great practice, I created one state isLogged that by default is false.
1 mutation Auth that contains as an arguments state and payload. I set the boolean isLogged equal to the payload value.
In Actions I created two actions login & logout, the only difference is that in login Auth is set to true.
In Getters I return the boolean isLogged,
In userAuth I created two methods and dispatch the mutations.
1 computed property which stays for v-if condition.
Lastly in the app I copy the same computed property that I wrote in the userAuth, which is the getter.
With one check using v-if=Auth the Auth is ready.
Now when you click login you can see the other information, and if you logout it's hidden.