Separated Module Auth
Created separated module Auth. I didn't namespace it, because the userId is global, in my way I won't need to change anything else to make it work. Just removing the getters from the main store to the auth/getters.js and the state from the main store to auth/index.js. In this way I get them. Importing the AuthModule in the store .