Signup generally works
Created signup action, that fetch data from FireBase with the give project API_Key. Converted into JSON format - the email, the password and the secureToken must be true. New mutation added setUser(), that takes the token, userId and tokenExpiration. In Index.js I make the by default to be null. In the signup method I commit the mutation. In UserAuth component In the submitForm method, I dispatch only the signup method, with it's email and password as a state. Now working on giving the user some feedback.