Authorization complete
Now pages that want authorization are complete, by creating new getter - that takes the token. In messages actions, I get the token from the rootGetters and save it in constant. Now the pages that require authorization are the one messages, where the fetch messages method is. At the end of the url I added ?auth= this is a FireBase thing and + the token I got. In Pets actions is the same, for registering Pets. Now if you are logged in you can visit your messages and you can register your pet.