Updating the UI
Based on authorization, update the UI with getters. How I can tell somebody that is not logged in, by the tokens. So I created new getters that I convert to boolean and named it IsAuthenticated. In PetsList I create new computed property isLoggedIn and I check if the user is not logged first. If that is the case I provided second base-button (link) that goes to the auth page. In the header I created the same computed property and used it with v-if again. If is the user is not logged in, it can't see the requests but there I provided again new button/router-link to go to the auth page. Now with the logout is pretty simple. I created new action logout. I commit the mutation setUser and passed the payload data but set the tokenId, userId and tokenExpiration back to null. After the commit in the header component I created method in which I dispatch the action for logout and set this method to the logout button with the short form of v-on.