Navigation guards.
In the router.js I defined an navigation guards that check if the path requires auth and authentication or the path don't require authentic. defining requiresAuth: variable. In pet's list make the user experience better by improving the url. In the header and in the UserAuth component using the replace method to return to the pet's page after doing some action.