AutoLogin
Making one method auth and saving code, for signup and login methods. In const I saved the mode payload. Create a variable that saves the url for login. Checking if the mode is signup, if it's then the variable is using the sign up url. Saving in the localStorage the tokenId and the localId. Creating new tryLogin method that get's the tokenId and the userId from the local storage and checks if they exist in , if they exist then I commit the setUser with it's payload. I run tryLogin method in the app component using created lifecycle hook.