AutoLogout
Removing the item from the localStorage and checking if the expiration token is expired . In actions.js the method tryLogin setTime out and dispatch the autoLogout method. Before that I created a variable named timer. Main point of it is to clear the interval when I need it like in the logout method. Created new getters didAutoLogout and new mutation that sets it to true. In App component I return the didAutoLogout getter in computed property. Created a watcher that looks for current value & old value. If I have logout on page that requires authentication, I will be redirected to pets list.